refactor(kernel): delete wal git accept path
