--- packages/kernel/src/store/wal-git-materializer.ts
@@ -1,298 +1,0 @@
-import { serializePersistedCanonicalEvent } from "../domain/doc-sync.contract.ts";
-import { consumeKnownError } from "../error-consumption.ts";
-import { serializeEventHead, type EventHead } from "../domain/write-chain.contract.ts";
-import { sha256Text } from "../integrity/stable-hash.ts";
-import type { HarnessLayoutInput } from "../layout/index.ts";
-import {
-  contentObjectRelativePath,
-  eventObjectRelativePath,
-  type LedgerLayoutState,
-  type LedgerObjectLayout,
-} from "../layout/ledger-object-layout.ts";
-import { ledgerGitPath, resolveLedgerGitLayout } from "./ledger-git-layout.ts";
-import { localGitObjectRefStore, localGitWorktreeSettlement } from "./local-version-control-system.ts";
-import {
-  canonicalDocumentClaims,
-  canonicalDocumentMode,
-  canonicalDocumentRetirements,
-  TaskEventStoreError,
-  type CanonicalEventStore,
-  type EventPublicationKillpoint,
-} from "./task-event-store.ts";
-import type { WalEventRecord, WalMaterializationSource } from "./wal-event-log.ts";
-
-export interface WalGitMaterializerOptions {
-  readonly rootInput?: HarnessLayoutInput;
-  readonly rootDir?: string;
-  readonly authoredBranch?: string;
-  readonly killpoint?: (point: EventPublicationKillpoint) => void;
-  readonly withAppendFence?: <T>(operation: () => T) => T;
-  /** Keep immutable event/content objects in Git while excluding them from the active worktree. */
-  readonly compactWorktree?: boolean;
-  /** Layout already verified for the expected Git cut by the materialization host. */
-  readonly layout: LedgerLayoutState;
-}
-
-export interface WalGitMaterializationReceipt {
-  readonly commitSha: string;
-  readonly head: EventHead | null;
-  readonly layout: LedgerLayoutState;
-}
-
-/** The authored branch moved independently of the canonical daemon cut. */
-export class WalMaterializerDivergedError extends TaskEventStoreError {
-  readonly canonicalSha: string;
-
-  constructor(repoRoot: string, authoredRef: string, canonicalSha: string) {
-    super(
-      "publication_indeterminate",
-      [
-        `authored ref ${authoredRef} diverged from the canonical event cut ${canonicalSha}.`,
-        `Repair with: git -C ${repoRoot} reset ${canonicalSha}`,
-        "then retry materialization.",
-      ].join(" "),
-    );
-    this.name = "WalMaterializerDivergedError";
-    this.canonicalSha = canonicalSha;
-  }
-}
-
-/** An authored ref advanced compatibly while a WAL commit was being prepared. */
-export class WalMaterializerRefChangedError extends TaskEventStoreError {
-  constructor(authoredRef: string) {
-    super("publication_indeterminate", `${authoredRef} advanced while finalizing WAL materialization; retry the cut`);
-    this.name = "WalMaterializerRefChangedError";
-  }
-}
-
-export function flushWalToGit(
-  wal: WalMaterializationSource,
-  git: CanonicalEventStore,
-  options: WalGitMaterializerOptions,
-): WalGitMaterializationReceipt {
-  const records = wal.records();
-  if (records.length === 0)
-    return {
-      commitSha: git.currentCommit().sha,
-      head: git.readHead(),
-      layout: options.layout,
-    };
-  const input = options.rootInput ?? options.rootDir;
-  if (input === undefined) throw new Error("canonical event store requires rootInput or rootDir");
-  const ledger = resolveLedgerGitLayout(input);
-  cleanFlushRefs(ledger.rootDir);
-  const through = git.readHead()?.revision ?? 0;
-  for (const record of records.filter((candidate) => candidate.revision <= through)) {
-    const durable = git.readEvent(record.opId);
-    if (durable === null || canonicalBytes(durable) !== canonicalBytes(record.event))
-      throw new TaskEventStoreError(
-        "invalid_store",
-        `WAL revision ${record.revision} differs from its materialized Git event`,
-      );
-  }
-  const pending = records.filter((record) => record.revision > through);
-  const settlementRecords = pending.length === 0 ? records : pending;
-  const layoutState = options.layout;
-  if (layoutState === "mixed")
-    throw new TaskEventStoreError("invalid_store", "cannot flush WAL into a mixed Git layout");
-  const files = batchFiles(wal, ledger, settlementRecords, layoutState);
-  const documentTargets = new Set(
-    settlementRecords.flatMap((record) => [
-      ...canonicalDocumentClaims(record.event).map((claim) => ledgerGitPath(ledger, claim.path)),
-      ...canonicalDocumentRetirements(record.event).map((retirement) => ledgerGitPath(ledger, retirement.path)),
-    ]),
-  );
-  const documentFiles = files.filter((file) => documentTargets.has("delete" in file ? file.delete : file.target));
-  const immutableFiles = options.compactWorktree
-      ? files.filter((file) => immutableLedgerObject(ledger, "delete" in file ? file.delete : file.target))
-      : [],
-    immutableTargets = new Set(immutableFiles.map((file) => ("delete" in file ? file.delete : file.target))),
-    durableFiles = files.filter((file) => {
-      const target = "delete" in file ? file.delete : file.target;
-      return !documentTargets.has(target) && !immutableTargets.has(target);
-    });
-  if (pending.length === 0) {
-    localGitWorktreeSettlement.settle(ledger.rootDir, durableFiles);
-    localGitWorktreeSettlement.index(ledger.rootDir, documentFiles);
-    localGitWorktreeSettlement.index(ledger.rootDir, immutableFiles, true);
-    return { commitSha: git.currentCommit().sha, head: git.readHead(), layout: layoutState };
-  }
-  for (const [index, record] of pending.entries())
-    if (record.revision !== through + index + 1)
-      throw new TaskEventStoreError(
-        "revision_conflict",
-        `WAL revision ${record.revision} is not contiguous after Git revision ${through}`,
-      );
-  const last = pending.at(-1)!;
-  const parent = git.currentCommit().sha;
-  const branch = options.authoredBranch ?? localGitObjectRefStore.currentBranch(ledger.rootDir);
-  if (!branch)
-    throw new TaskEventStoreError(
-      "publication_indeterminate",
-      "authored branch is detached; register a default branch before flushing WAL",
-    );
-  const authoredRef = `refs/heads/${branch}`;
-  const readLiveRefs = (): Map<string, string> =>
-      new Map(
-        localGitObjectRefStore
-          .listRefs(ledger.rootDir, ["refs/ha/canonical", authoredRef])
-          .trim()
-          .split(/\r?\n/u)
-          .filter(Boolean)
-          .map((line) => {
-            const [ref, sha] = line.split(" ");
-            return [ref!, sha!] as const;
-          }),
-      ),
-    liveRefs = readLiveRefs();
-  const canonicalParent = liveRefs.get("refs/ha/canonical");
-  const authoredParent = liveRefs.get(authoredRef);
-  if (canonicalParent !== parent || authoredParent === undefined)
-    throw new TaskEventStoreError(
-      "publication_indeterminate",
-      "Git refs changed while preparing the WAL materialization batch",
-    );
-  if (authoredParent !== parent && !localGitObjectRefStore.isAncestor(ledger.rootDir, parent, authoredParent))
-    throw new WalMaterializerDivergedError(ledger.rootDir, authoredRef, parent);
-  const flushRef = `refs/ha-wal-flush/${process.pid}-${Date.now()}`;
-  const messageText = `harness WAL flush ${pending[0]!.revision}-${last.revision}`;
-  const timestamp = Math.floor(Date.parse(last.event.occurredAt) / 1_000);
-  let inputText = `commit ${flushRef}\nmark :1\ncommitter Harness WAL <harness-wal@local.invalid> ${timestamp} +0000\ndata ${Buffer.byteLength(messageText)}\n${messageText}\nfrom ${authoredParent}\n`;
-  for (const file of files)
-    inputText +=
-      "delete" in file
-        ? `D ${file.delete}\n`
-        : `M ${file.mode ?? "100644"} inline ${file.target}\ndata ${Buffer.byteLength(file.body)}\n${file.body}\n`;
-  inputText += "\nget-mark :1\ndone\n";
-  const imported =
-    localGitObjectRefStore.importCommit(ledger.rootDir, inputText).toString("utf8").trim().split("\n").at(-1) ?? "";
-  if (!/^[0-9a-f]{40}$/u.test(imported))
-    throw new TaskEventStoreError("publication_indeterminate", "WAL Git flush returned no commit");
-  options.killpoint?.("after_git_commit");
-  const finalize = (): void =>
-    localGitObjectRefStore.updateRefs(
-      ledger.rootDir,
-      `start\nupdate refs/ha/canonical ${imported} ${parent}\nupdate ${authoredRef} ${imported} ${authoredParent}\nprepare\ncommit\n`,
-    );
-  try {
-    if (options.withAppendFence) options.withAppendFence(finalize);
-    else finalize();
-    options.killpoint?.("after_git_ref_update");
-    localGitWorktreeSettlement.settle(ledger.rootDir, durableFiles);
-    localGitWorktreeSettlement.index(ledger.rootDir, documentFiles);
-    localGitWorktreeSettlement.index(ledger.rootDir, immutableFiles, true);
-  } catch (error) {
-    try {
-      const currentRefs = readLiveRefs(),
-        currentCanonical = currentRefs.get("refs/ha/canonical"),
-        currentAuthored = currentRefs.get(authoredRef);
-      if (currentCanonical === parent && currentAuthored !== undefined && currentAuthored !== authoredParent) {
-        consumeKnownError(error);
-        if (!localGitObjectRefStore.isAncestor(ledger.rootDir, parent, currentAuthored))
-          throw new WalMaterializerDivergedError(ledger.rootDir, authoredRef, parent);
-        throw new WalMaterializerRefChangedError(authoredRef);
-      }
-    } catch (classificationError) {
-      if (
-        classificationError instanceof WalMaterializerDivergedError ||
-        classificationError instanceof WalMaterializerRefChangedError
-      )
-        throw classificationError;
-      consumeKnownError(classificationError);
-    }
-    throw error;
-  } finally {
-    localGitObjectRefStore.deleteRef(ledger.rootDir, flushRef);
-  }
-  return {
-    commitSha: imported,
-    head: {
-      revision: last.revision,
-      opId: last.opId,
-      eventDigest: last.eventDigest,
-    },
-    layout: layoutState,
-  };
-}
-
-function immutableLedgerObject(ledger: ReturnType<typeof resolveLedgerGitLayout>, target: string): boolean {
-  const events = ledgerGitPath(ledger, "events/"),
-    objects = ledgerGitPath(ledger, "objects/sha256/");
-  return (target.startsWith(events) && target !== `${events}head.json`) || target.startsWith(objects);
-}
-
-function batchFiles(
-  wal: WalMaterializationSource,
-  ledger: ReturnType<typeof resolveLedgerGitLayout>,
-  records: readonly WalEventRecord[],
-  layout: LedgerObjectLayout,
-): readonly (
-  | { readonly target: string; readonly body: string; readonly mode?: "100644" | "120000" }
-  | { readonly delete: string }
-)[] {
-  const files = new Map<
-    string,
-    | { readonly target: string; readonly body: string; readonly mode?: "100644" | "120000" }
-    | { readonly delete: string }
-  >();
-  const add = (target: string, body: string, mode: "100644" | "120000" = "100644", replace = false): void => {
-    const existing = files.get(target);
-    if (
-      !replace &&
-      existing !== undefined &&
-      ("delete" in existing || existing.body !== body || existing.mode !== mode)
-    )
-      throw new TaskEventStoreError("op_conflict", `WAL batch names different bytes at ${target}`);
-    files.set(target, { target, body, mode });
-  };
-  const remove = (target: string): void => {
-    files.set(target, { delete: target });
-  };
-  for (const record of records) {
-    add(ledgerGitPath(ledger, eventObjectRelativePath(record.opId, layout)), canonicalBytes(record.event));
-    for (const blob of record.blobs)
-      add(
-        ledgerGitPath(ledger, contentObjectRelativePath(blob.sha256, layout)),
-        Buffer.from(requiredWalBlob(wal, blob.sha256)).toString("utf8"),
-      );
-    for (const retirement of canonicalDocumentRetirements(record.event)) remove(ledgerGitPath(ledger, retirement.path));
-    for (const claim of canonicalDocumentClaims(record.event))
-      add(
-        ledgerGitPath(ledger, claim.path),
-        Buffer.from(requiredWalBlob(wal, claim.sha256)).toString("utf8"),
-        canonicalDocumentMode(record.event, claim.path),
-        true,
-      );
-  }
-  const last = records.at(-1)!;
-  add(
-    ledgerGitPath(ledger, "events/head.json"),
-    serializeEventHead({
-      revision: last.revision,
-      opId: last.opId,
-      eventDigest: `sha256:${sha256Text(canonicalBytes(last.event))}`,
-    }),
-    "100644",
-    true,
-  );
-  return [...files.values()];
-}
-
-function requiredWalBlob(wal: WalMaterializationSource, sha256: string): Uint8Array {
-  const bytes = wal.readContentBlob(sha256);
-  if (bytes === null) throw new TaskEventStoreError("invalid_store", `WAL content object ${sha256} is missing`);
-  if (sha256Text(Buffer.from(bytes).toString("utf8")) !== sha256)
-    throw new TaskEventStoreError("invalid_store", `WAL content object ${sha256} is corrupt`);
-  return bytes;
-}
-
-function cleanFlushRefs(repoRoot: string): void {
-  for (const line of localGitObjectRefStore.listRefs(repoRoot, ["refs/ha-wal-flush/"]).trim().split(/\r?\n/u)) {
-    const ref = line.split(" ")[0];
-    if (ref) localGitObjectRefStore.deleteRef(repoRoot, ref);
-  }
-}
-
-function canonicalBytes(event: WalEventRecord["event"]): string {
-  return serializePersistedCanonicalEvent(event);
-}
