+++ packages/kernel/test/store/task-projection.test.ts
@@ -12,7 +12,6 @@ import {
   type CanonicalEventStore,
   type CanonicalWriteBundle,
 } from "../../src/store/task-event-store.ts";
-import { makeWalShadowEventStore } from "../../src/store/wal-shadow-event-store.ts";
 import { localGitObjectRefStore } from "../../src/store/local-version-control-system.ts";
 import { taskLifecycleWritePlan } from "../../src/domain/task-lifecycle-publication.ts";
 import { compileEntityUpsert } from "../../src/domain/entity-event.ts";
@@ -431,28 +430,6 @@ test("migration-sized catch-up reduces 5k events in bounded projection transacti
     const receipt = projection.catchUp!();
     assert.equal(receipt.watermark, count);
     assert.deepEqual(receipt.metrics, { sqliteTransactions: 3, reducedItems: count, maxBatchItems: 4_096 });
-    projection.close();
-  });
-});
-
-test("WAL-shadow cold projection preserves Git batch content prefetch", async (t) => {
-  await withTempStoreAsync(async (rootDir) => {
-    initRepo(rootDir);
-    const writer = makeTaskEventStore({ repoId: "wal-shadow-batch-prefetch", rootDir }),
-      count = 128;
-    for (let revision = 1; revision <= count; revision += 1) writer.append(batchDocumentBundle(writer, revision));
-    const reader = makeWalShadowEventStore({ repoId: "wal-shadow-batch-prefetch", rootDir, walFlushMs: 60_000 }),
-      projection = makeTaskProjection({ rootDir, eventStore: reader, catchUpLimit: 64 }),
-      before = localGitObjectRefStore.processCount();
-    const rebuilt = projection.rebuild(),
-      processes = localGitObjectRefStore.processCount() - before;
-    t.diagnostic(JSON.stringify({ events: count, gitProcesses: processes, watermark: rebuilt.watermark }));
-    assert.equal(rebuilt.watermark, count);
-    assert.equal(
-      processes <= 7,
-      true,
-      `WAL-shadow cold projection opened ${processes} Git processes for ${count} claimed blobs across two batches`,
-    );
     projection.close();
   });
 });
